Script to fetch pinyin based on raw txt file
Python3 script to fetch pinyin based on raw txt file. Results will be saved in the same dir with same filename with suffix "dict_".

format of the txt file should be 
"词语 freq # comments"
"词语 freq" or 
"词语"

However only only the fist two parts (i.e.m 词语 and freq) will be retained in the result.
The generated result file can be imported to ibus-libpinyin as user dicts directly.

Author: Kevin Suo <suokunlong@126.com>
This script is written specially for libpinyin, but may be used for any other suitable purposes.

Licensed under GNU Public License v3 or above.